release: update and add another image

+++ release.sh
@@ -11,10 +11,11 @@ have_image_2=0
 have_image_3=0
 have_image_4=0
 have_image_5=0
+have_image_6=0
 
 usage()
 {
-        echo "Usage: $0 -v version -i img -f img -g img -h img -j img"
+        echo "Usage: $0 -v version -i img -f img -g img -h img -j img -k img"
 }
 
 args=$(getopt -o v:i:f:g:h:j: -- "$@")
@@ -49,6 +50,11 @@ do
 	-j)
 		RELEASE_IMAGE_5=$2
 		have_image_5=1
+		shift
+		;;
+	-k)
+		RELEASE_IMAGE_6=$2
+		have_image_6=1
 		shift
 		;;
         -v)
@@ -95,7 +101,11 @@ if [ ! "$have_image_5" -gt 0 ] ; then
 	usage
 	exit 1
 fi
-
+if [ ! "$have_image_6" -gt 0 ] ; then
+	echo "image missing"
+	usage
+	exit 1
+fi
 
 # Do we have a desired version number?
 if [ "$have_version" -gt 0 ] ; then
@@ -162,6 +172,18 @@ if [ ! "$filesize" -eq "$reference_filesize" ] ; then
 	echo "filesize of release image is wrong"
 	exit 1
 fi
+filesize=$(wc -c <"${RELEASE_IMAGE_5}")
+reference_filesize=4194304
+if [ ! "$filesize" -eq "$reference_filesize" ] ; then
+	echo "filesize of release image is wrong"
+	exit 1
+fi
+filesize=$(wc -c <"${RELEASE_IMAGE_6}")
+reference_filesize=4194304
+if [ ! "$filesize" -eq "$reference_filesize" ] ; then
+	echo "filesize of release image is wrong"
+	exit 1
+fi
 
 RELEASE_DIR="skulls-${version}"
 rm -rf "$RELEASE_DIR"
@@ -173,6 +195,7 @@ cp "${RELEASE_IMAGE_2}" "$RELEASE_DIR"
 cp "${RELEASE_IMAGE_3}" "$RELEASE_DIR"
 cp "${RELEASE_IMAGE_4}" "$RELEASE_DIR"
 cp "${RELEASE_IMAGE_5}" "$RELEASE_DIR"
+cp "${RELEASE_IMAGE_6}" "$RELEASE_DIR"
 
 RELEASE_IMAGE_FILE=$(basename "${RELEASE_IMAGE}")
 sha256sum ${RELEASE_DIR}/${RELEASE_IMAGE_FILE} > "${RELEASE_DIR}/${RELEASE_IMAGE_FILE}.sha256"
@@ -184,6 +207,8 @@ RELEASE_IMAGE_FILE_4=$(basename "${RELEASE_IMAGE_4}")
 sha256sum ${RELEASE_DIR}/${RELEASE_IMAGE_FILE_4} > "${RELEASE_DIR}/${RELEASE_IMAGE_FILE_4}.sha256"
 RELEASE_IMAGE_FILE_5=$(basename "${RELEASE_IMAGE_5}")
 sha256sum ${RELEASE_DIR}/${RELEASE_IMAGE_FILE_5} > "${RELEASE_DIR}/${RELEASE_IMAGE_FILE_5}.sha256"
+RELEASE_IMAGE_FILE_6=$(basename "${RELEASE_IMAGE_6}")
+sha256sum ${RELEASE_DIR}/${RELEASE_IMAGE_FILE_6} > "${RELEASE_DIR}/${RELEASE_IMAGE_FILE_6}.sha256"
 
 # copy in device independent stuff
 cp SOURCE.md "$RELEASE_DIR"
